File names changed to lowercase versions General cleanup and linkage of classes
